--- a/chunk_downloader.py
+++ b/chunk_downloader.py
@@ -70,8 +70,7 @@ class SnowflakeChunkDownloader(object):
                                            self._chunk_size)
 
         for idx, chunk in enumerate(chunks):
-            logger.info(u"queued chunk: url=%s, rowCount=%s",
-                        chunk[u'url'], chunk[u'rowCount'])
+            logger.info(u"queued chunk %d: rowCount=%s" % (idx, chunk[u'rowCount']))
             self._chunks[idx] = SnowflakeChunk(
                 url=chunk[u'url'],
                 result_data=None,