--- a/chunk_downloader.py
+++ b/chunk_downloader.py
@@ -96,8 +96,7 @@ class SnowflakeChunkDownloader(object):
     def __init__(self, chunks, connection, cursor, qrmk, chunk_headers,
                  prefetch_slots=DEFAULT_CLIENT_RESULT_PREFETCH_SLOTS,
                  prefetch_threads=DEFAULT_CLIENT_RESULT_PREFETCH_THREADS,
-                 use_ijson=False, row_filter=None):
-        self._row_filter = row_filter
+                 use_ijson=False):
         self._pre_init(chunks, connection, cursor, qrmk, chunk_headers,
                        prefetch_slots=prefetch_slots,
                        prefetch_threads=prefetch_threads,
@@ -263,5 +262,5 @@ class SnowflakeChunkDownloader(object):
         data = self._connection.rest.fetch(u'get', url, headers,
             timeouts=timeouts, is_raw_binary=True,
             is_raw_binary_iterator=False, use_ijson=self._use_ijson)
-        return iter([self._row_filter(x) if x is not None else None
+        return iter([self._cursor.row_to_python(x) if x is not None else None
                      for x in data])