Micro optimizations for hot path code used to convert JSON raw data to native python data structures.  Most of the optimizations come from reducing function call overhead (by way of closures) and by more efficient use of `datetime.datetime` factories like `fromtimestamp()`. 

EDIT: Benchmarks.

In a micro benchmark with heavy datetime data sets and S3 downloads cached I see the following numbers..

Stock - Py3.5: 21.9s
Stock - Py3.6: 17.0s
This PR - Py3.5: 12.7s (72% faster)
This PR - Py3.6: 9.5s (79% faster)

These number are for 500k rows and include startup times.  The actual row processing speeds are over 100% faster.