@smtakeda 2017-03-16
Timestamp match issue is very likely due to http://bugs.python.org/issue23517
We use Python 3.4.3 and 2.7.10 internally. Here is a test code:

```
from datetime import datetime

ts='58510.989102000'
ret = datetime.utcfromtimestamp(float(ts)).time()
print('   ' + str(ts))
print(str(ret))
```
The results are:
```
>python2.6 test_time_precision.py
   58510.989102000
16:15:10.989102
>python2.7 test_time_precision.py  
   58510.989102000
16:15:10.989102
>python3.3 test_time_precision.py 
   58510.989102000
16:15:10.989101
>python3.4 test_time_precision.py 
   58510.989102000
16:15:10.989101
>python3.5 test_time_precision.py 
   58510.989102000
16:15:10.989102
>python3.6 test_time_precision.py 
   58510.989102000
16:15:10.989102
```