@smtakeda 2017-03-16
I would like a have Python 3.4.3 special converter to workaround http://bugs.python.org/issue23517.