@smtakeda 2017-03-16
3.5.0 has the same issue. 3.6.0 has the fix.

https://docs.python.org/3.4/whatsnew/changelog.html#python-3-4-4
https://docs.python.org/3.5/whatsnew/changelog.html#python-3-5-1-final
https://docs.python.org/3.6/whatsnew/changelog.html#python-3-6-0